base_list: Init commit
Adding a testcase for perf list. It is almost finished, but some
machine specific checks should be added there:

  * do we have hv_24x7 events on POWER?

  * do we have uncore events on CPUs that should have them?

  * etc.

This set of testcases should check also for presence of some
special events based on machine type and architecture.

Note: The presence of various event type groups like 'hw', 'sw',
'cache' etc. is not checked now. Those filters are tested for
excluding the other groups of event only. We should test which
groups are available, but it is also architecture-related, so
let's have it as TODO.