base_stat: Added test for Intel uncore events
In case there are any Intel uncore events available, this test uses
them with perf-stat. Otherwise the test is skipped.

The fact *whether* uncore events should be supported on a particular
machine should be tested in base_list testcases once.

Zero values are accepted as correct results too. This thing should
be rethought once more. But nowadays it is pretty usual, that an
uncore event is shown in perf-list's output, but when used it gives
"<not supported> string. This test should catch such situations.