base_probe: added test_probe_syntax.sh for checking probe syntax forms
Like the '-L' option, also the probes can be specified using various
formats. This test is added to check whether they are accepted.

Since the 'SRC:ALN' format is not easy to be used with kernel probes,
this particular thing is tested with a userspace probe. Some example
code has been added for it.

The example code contains very long symbol names in order to check,
whether the strings over 64 bytes can be accepted too. They are not
at the moment, but after fixing that thing, this test can be used
for testing it.

There is also issue with listing that should be investigated further.
When listing ALNs in kernel source code, it seems to work, but when
you want to add a probe at a line where a function begins, it is not
possible, despite that the line is white in the output and has a line
number.

The example can be seen in vfs_read:

$ perf probe -L fs/read_write.c | grep 'ssize_t vfs_read'

    440  ssize_t vfs_read(struct file *file, char __user *buf, size_t count, loff_t *pos)
    826  ssize_t vfs_readv(struct file *file, const struct iovec __user *vec,

The following lines after the 440 look like:

    440  ssize_t vfs_read(struct file *file, char __user *buf, size_t count, loff_t *pos)
    441  {
         	ssize_t ret;

    444  	if (!(file->f_mode & FMODE_READ))
    445  		return -EBADF;
    446  	if (!(file->f_mode & FMODE_CAN_READ))
    447  		return -EINVAL;
    448  	if (unlikely(!access_ok(VERIFY_WRITE, buf, count)))
    449  		return -EFAULT;

But when trying to put a probe at the line 440, it fails:

$ perf probe fs/read_write.c:440
Failed to find scope of probe point.
  Error: Failed to add events.

Adding it to line 444 works well (441 too).

$ perf probe fs/read_write.c:444
Added new event:
  probe:vfs_read       (on @fs/read_write.c:444)

You can now use it in all perf tools, such as:

	perf record -e probe:vfs_read -aR sleep 1

Due to that, it is not straightforward to add a guessed ALN in kernel
code.

Another strange thing is that the userspace code does not support ALNs.
The following command ends with error:

$ perf probe -x examples/test -L test.c

Maybe it is a bug, maybe just an unimplemented feature.