base_probe: a first part of test_listing.sh added
This is a testcase for checking various listing options like the
following:

  -L	for line listing
  -F	for function listing
  -V	for variable listing

These options should be tested both with kernel and userspace
targets.

The '-F' option is covered somehow already, the rest needs to be
covered in the upcoming commits. For userspace, it is pretty easy
to check, whether the right functions are listed or not. In case
of kernel it is compared against kallsyms by special perl based
checker (check_kallsyms_vs_probes.pl).

The checker takes the /proc/kallsyms file and the output of 'perf
probe -F' and counts how many of the probe points are present in
the /proc/kallsyms file. First testing shows that 96% of the syms
listed by 'perf probe -F' is present in /proc/kallsyms. That is
not bad for now.

Maybe it deserves a bit deeper checking, but this should be good
enough for now. In case the rate is lower than 85%, the test fails.