base_stat: adding test_tracepoints_definition.sh
Some of the tracepoint events cause warnings when they are used as
events i.e. by perf-stat. This test goes through all the tracepoint
events offered by perf-list and checks each one for such warnings.

Some events are fiving "<not supported>" value despite being listed
by perf list. The tests shows that too.

Since it takes around 35 secs and it is not that important, there
is an option which can disable it. Maybe we should consider disabling
that by default.