base_probe: test_advanced.sh added
This testcase tests more advanced probing. Since the probes can
carry values of variables, registers and return values and perf
record can record them, we need to check, whether it works well
and whether the values are correct.

The testcase actually fails with the latest kernel - the values
are shifted somehow, so instead of:

a=2
a=3
a=4
...
a=19

we get:

a=32766
a=32582
a=3
a=4
...
a=17

so the last value is lost and the first one as well.  When the
same value is tracked through register (%ax), it seems to work
correctly.

The test needs to be extended to check registers too, but that
is very arch specific. Return values need to be tested too and
various other data types.