base_buildid: Init commit
Adding some tests for perf buildid-list and buildid-cache (future)
commands. This commit adds a testcase for buildid-list only and it
checks, whether the buildids grabbed from perf.data file match the
real files. A new perl checker has been added for verifying that.

The tool has also and option ("--kernel") for printing the buildid
of the running kernel.  That functionality has been broken in perf
some time ago,  but a patch for it has been sent,  so the testcase
will be passing soon hopefully.

TODO related to this:

  * add some testcases for perf buildid-cache tool
    * check if the buildid-cache operations do what they should
    * check the consistency of the ~/.debug/* structure against
    the real executables