test_driver.sh: update what to run where
Basically, the tests should be supported everywhere, except of some
arch-specific (intel-uncore, powerpc-hv_24x7, ...). The reality is,
that x86 is lead arch,  so it is always some steps further and some
features work only there; and some fixes are needed in order to run
on the other archs. That's why the mechanism is there, but the goal
is to remove it completely, having everything passing (or correctly
skipping).