base_report: stderr checking more tolerant
Although stderr should be empty when calling perf-report, sometimes
it produces some error messages, that might be safely ignored. This
patch introduces a stderr checker with whitelisting mechanism. When
a stderr file is empty or the only lines there match to a regexp of
the whitelist, it is considered as PASS. Otherwise it means a FAIL.

The functionality is similar to check_all_lines_matched.pl,  but in
the case of whitelisting, it is more convenient to have a file than
passing the arguments through the command line.

Which error messages can be whitelisted?

I.e. missing debuginfo, since perf isn't responsible for having it.