general: Adding a new test-driver written in C
In order to integrate this testsuite we need a driver that would run
the whole suite, but being written in C.

That makes the integration much easier.

The C driver (run_all.c) does not hardcode any lists of what should
be run and where.  It enters all the base_* dirs and runs setup.sh,
all the test_* executables and finally the cleanup.sh.  If setup.sh
is missing, it is ignored, but other errors are considered.

Due to that base_skeleton needs to be renamed a bit, since it isn't
a regular test.