base_annotate: Move logs and temporary files outside the tree
When PERFSUITE_RUN_DIR environment variable is set, all logs and auxiliary
files that are generated (binaries, perf.data) by perf-annotate tests will
be moved to $PERFSUITE_RUN_DIR. A tree structure is created there.

When PERFSUITE_RUN_DIR environment variable is NOT set, everything goes to
current dir.  It is worth of a note that cleanup.sh has not changed, since
it would be risky to delete anything in a path specified by a variable. It
would need checks to avoid `find / -name \*.log | xargs rm` etc... The log
dir must be taken care by the user manualy later.