base_trace: overload test added
There was a bug in perf-trace which caused that system-wide tracing
was falling into an endless loop.  The following command should run
the system-wide tracing for three seconds and then finish:

perf trace -a -- sleep 3

With the mentioned bug it did not finish. The testcase added by the
commit should cover this bug.

Some other heavy testcases are expected to be added later.