base_probe/test_exact_counts.sh specify the columns in perf report
This commit reacts to kernel's d49dadea78624353d1df660efb49f187bd5c5971.

Since the default columns changed here, we have to specify exactly what
we want to display.

Before:
   100.00%         65535  (4005d8)
   100.00%           997  (4005d0)
   100.00%           103  (4005c8)
   100.00%             3  (4005c0)
   100.00%             2  (4005b8)
   100.00%             1  (4005b0)

After:
   100.00%         65535  exact_counts  exact_counts   [.] f_65535x
   100.00%           997  exact_counts  exact_counts   [.] f_997x
   100.00%           103  exact_counts  exact_counts   [.] f_103x
   100.00%             3  exact_counts  exact_counts   [.] f_3x
   100.00%             2  exact_counts  exact_counts   [.] f_2x
   100.00%             1  exact_counts  exact_counts   [.] f_1x

In order to check the matching of sample counts to functions, we need
to display symbols. The comm and dso is nice to have there too.