common: few env detection functions added to init.sh
We will need them later. When testing kernel/perf support of various
special CPU features, it would be nice to be able not only to skip
the particular testcase if the feature is not found, but also to fail
in case the feature should be supported and the test finds out that
it is not. For example, Intel uncore or Intel RAPL work only on some
specific models of Intel CPUs, and none of these features is supported
on virtual guests. There will be some AMD specific features in future
as well.