base_stat: uncore and RAPL support expectation introduced
The test_intel_uncore.sh now detects whether uncore is expected to be
supported on current CPU model. In case it is expected but not supported,
the testcase fails. In case it is not supported and not expected, it is
just skipped as before.

The expectation is based on a function in settings.sh. A similar function
has been added for RAPL.