base_stat: added tests for excluding {kernel|userspace} feature
The events can be specified with various modifiers, like "cycles:XY".
Two of them, "k" and "u" are there to specify, whether perf should
measure what a process spends in kernel, in userspace or in both. By
default, perf counts both.

-e cycles
	counts both kernel and userspace activity of the process
-e cycles:u
	counts only when the process does userspace
-e cycles:k
	counts only when the process spends time in kernel (e.g. IO)
-e cycles:ku
	the same as "cycles"

It makes sense, that cycles:u + cycles:k should be equal to cycles:ku.
A test for this is introduced with this commit.

We do it only for hw and hwcache events. The second ones are disabled,
because perf does not print the modifiers along with the event names.
This might be a bug, we'll see how the discussion on perf-users-list
will show.

The following hw events are disabled from this test too:
ref-cycles
stalled-cycles-backend
stalled-cycles-frontend

It seems that it does not make sense to test these. However, it should
be double-checked in future (TODO, FIXME).