base_record/test_overhead.sh: new stress test added for perf record
                             * * * * *

NOTE: It is pretty possible, that the test will need some more tweaking.
It is still in an alpha version.

!!! WARNING: Please note that the test might be really stressful and
some kernel/perf versions triggered a kernel crash. Do not run this
test on a production machine.

ALSO NOTE that the perf.data files will be very large, it depends on
the value of `nproc` on your machine, however, they might need gigabytes
of space.

If you do not like this test, you can disable it by running the suite
with 'PARAM_RECORD_OVERLOAD=n'.

                             * * * * *

This test runs some sample program nproc-times on background and tries
to profile system-wide using perf record.

We run the profiler three times, simple, with 'fp' call-graph and with
'dwarf' call-graph. Each profiling runs for 30 seconds, so it takes 1.5
min itself.

Each run of perf record produces a perf.data file which is processed
by perf report to check whether they contain some symbols and whether
they are generally sane.