base_stat/test_tracepoints_definition.sh: logging warnings
The test for tracepoint definition should verify, that there are no
e.g. syntax problems in the implementation of particular tracepoints.

Currently there are lots of issues that can be seen as warnings when
an event is used, e.g.:

$ perf stat -e kvm:kvm_arm_set_regset -- true
  Warning: [kvm:kvm_arm_set_regset] function sizeof not defined
  Warning: Error: expected type 5 but read 0

With this commit, a more detailed logging has been added, so we can
e.g. grep for a particular warning to see which tracepoints contain
it, etc.