common/check_ku_sum.pl: division-by-zero fixed; fail fails now
* When an event returns 0, it caused division by zero in "k+u=ku"
tests. This is because of the formula used for verification, which
is  (k + u - ku) / ku <> 0.003  --> this is because we want to be
a little bit benevolent and allow small relative diffs. When ku=0,
we divide by 0 which is of course wrong. After this fix, a basic
"k + u - ku <> 0" formula is used instead (fails in case k+u > 0
then).

* A missing "$passed = 0" statement in case of failure caused this
script to hide failures from outside.