base_list/test_powerpc_hv_24x7.sh: new testcase
This test tests correct listing of hv_24x7 events and their names.
These events are available on POWER8 LPAR machine only. On other
arches/machines it should be skipped.

A correct name is the following:

hv_24x7/HPM_SYNC_PCYC,domain=?,core=?/

An incorrect one is this:

hv_24x7/HPM_TC_2_CCYC__PHYS_CORE,core=?/

The domain should be specified as an argument, not being a part of
the name. The domain numbers are the following:

$ cat /sys/bus/event_source/devices/hv_24x7/interface/domains
1: Physical Chip
2: Physical Core
3: VCPU Home Core
4: VCPU Home Chip
5: VCPU Home Node
6: VCPU Remote Node