base_annotate/setup.sh: accept messages about lowering sample rate
Messages like the following one are often present in the stderr of
perf-record when recording the load for perf-annotate test.  Since
it is completely fine when the sample rate gets lowered, we should
not raise an alarm due to that.

Lowering default frequency rate to 1000.
Please consider tweaking /proc/sys/kernel/perf_event_max_sample_rate.