base_mem: introducing new testcase
The command perf-mem is a shortcut to perf-record and perf-report
called with some particular arguments leading to a memory-related
profile (stores and loads).  This commit introduces a basic tests
for this sub-tool.

It should use events cpu/mem-stores/ and cpu/mem-loads/.  In case
these events are not listed in perf list, the corresponding perts
of the tests are skipped.

We also need to check that perf does NOT use cycles (a bug).

TODO: We need some better example code to measure and also some a
bit more advanced event correctness checking.