base_buildid: actual fixing of non-invasive buildid-cache clean-up
Before, all the buildid data was stored in ~/.debug. This is the
case of all the remaining tests too, however, in this test, we need
to clean the cache and do other invasive changes there, thus this
particular test uses separate buildid-cache dir: ~/.debug-`date +%s`

However, before, this dir was always created, but not used.

After this fix, the buildid data is stored in ~/.debug-`date +%s`
and this dir is also properly deleted within cleanup.sh. Thus, this
test should no more spoil the default ~/.debug dir and, which is the
most important, it should not delete any user data from there.

In order to fix this, we had to do some more changes in there. The
setup.sh converted into the non-executable setup_src.sh which is to
be sourced in each test instead of being run manually. Thus, each
testcase (nowadays only one, but in future there might be several of
them) sets the things up itself.

All the temp .debug-`date +%s` dirs are collected in BUILDIDDIRS
file which is processed by cleanup.sh and this is how the temporary
dirs are cleaned and deleted after.

Also, since also cleanup.sh sources settings.sh, we needed to put
the buildid-cache settings into a separate file, which is sourced
only and only when the settings.sh is sourced by a test. When it is
sourced by cleanup.sh, sourcing of settings_cache.sh is omitted.

***

The mentioned changes are a bit hackish, which is not ideal, but
probably necessary in order to fix the behaviour.

The same approach is going to be used in perf-archive test in future.