base_mem: test disabled in case $PERFSUITE_RUN_DIR is set
The $PERFSUITE_RUN_DIR variable is here to redirect all the 'rw'
accesses to some separate directory in order not to write anything
into the tree.

Since `perf mem record` does not support the '-o' option such as
`perf record` does, the perf.data file is always written to the
current directory.

In this situation, we have to skip this test when the user asks
us for not writting into the tree.

When this gets fixed in kernel, this commit may be reverted and
corresponding '-o $CURRENT_TEST_DIR/perf.data' option added to
record (the same for report '-i ...').