base_archive: new test for perf-archive
The perf-archive should pack the buildid cache to a tarball which can
be copied along with perf.data file to some other machine on which it
can  be analysed then.  The buildid cache  contains copies of all the
binaries (DSOs) that were hit by a sample.

This test records some samples, verifies them,then it packs the cache
by perf-archive and verifies that the archive tarball really contains
all the necessary DSOs.