base_diff: introducing new testcase
The command perf-diff is a tool for comparing two profiles in two
perf.data files. It is useful for comparing multiple runs of the
same application under different load (our example) or for whole
system profiling.

This testcase tests the most basic functionality. Some other tests
need to be added in order to cover the remaining options.