base_record/test_evlist.sh: introduced new testcase for perf-evlist
The perf-evlist is a helper that shows all the events contained in
perf.data (all the events of which there are samples). It should
list events, be able to print the sampling frequencies or dump the
whole event configuration for each of them.

Currently covered:

  * evlisting frequencies (100, 200 and 1000)
  * evlisting various events
  * help message

To be done:

  * evlisting events from perf.data which contains multiple events
  * verbose mode check

We use hw events here, because they are the most basic ones that
are reliable to produce samples. If we sampled cpu-migrations, it
might not produce anything (too few event occurrences, only one
CPU, etc.) and that would make the results unstable.

However, since not all machines support hw sampling, the test can
be skipped, if there is nothing to measure. This is not great, so
I will add some events of other kinds that are reliable and should
be available everywhere... in future...