base_list/test_basic.sh: updated regexps to handle the named events
Since perf has come with event naming similar to the events we know
from OProfile, we need to accept them in the perf-list output.

The format is not very parser-friendly,  as we are used to in perf.
Currently it looks like the following:

Event group:
  event.name1
       [Some description]
  event.name2
       [Some very long description that breaks into
        two lines]
  event.name3
       [a = b]

It is pretty possible, that the description breaks into three lines
in future, however, that would really chew up the power of our reg.
expressions,  so currently, we expect lines containing \[.* or \].*
as the event description.