base_probe/test_exact_counts.sh: fixed compatibility with kernel v4.9
Previously, the userspace probes were named by an abbreviation of the
program name, such as 'probe_exact:something' for a probe placed into
the 'exact_counts' binary.

Now, perf uses the whole name, at least in this case, thus the probes
are now called 'probe_exact_counts:something'.  We have to be able to
deal with both ways.

I really do not like changes like this one. Unfortunately, the devels
of perf do that pretty often.