base_list/test_basic.sh: updated regexps to handle the named events #2
commit eb128698bc159392b3364dbdd40c6b472135338f
Author: Michael Petlan <mpetlan@redhat.com>
Date:   Tue Dec 13 17:14:53 2016 +0100

The above commit added support for the named PMU events and for their
description in perf-list output. However, as expected, the description
can be longer than two lines, thus we need to have a regexp accepting
the lines in the middle. THIS SUCKS, because such a regexp must allow
almost everything. The only rule that seems to apply there is that the
description lines in between start with exactly 8 whitespaces.

However, this is NOT GOOD ENOUGH and perf-list output needs some more
clever parser, which would be able to accept all the specialties perf
prints out AND on the other hand not to be satisfied with any possible
crap (which loose regexps usually accept).

This fixes the perf-list test failures (basic+pmu) on ppc and x86.