base_stat/test_{hw,hwcache}.sh: fail if PMU is expected but not works
The only architecture that is allowed not to support PMU is currently
s390x. Also virtual machines and various older CPUs do not support it
but the default is to expect a working PMU.

If the testsuite does not detect PMU (hardware events), the "hw" and
"hwcache" tests fail. Except s390x of course.

You may disable this feature by setting TEST_IGNORE_MISSING_PMU="y".