base_trace: check correctness of the timestamps in trace output
It seems that sometimes the timestamps are bogus, such as:

     0.226 (126.564 ms): ...
   126.790 ( 0.011 ms): ...
18446162601971.988 ( 0.000 ms): ... <--- oh, my dear!
   126.815 (597.594 ms): ...
   724.409 ( 0.005 ms): ...

So this patch adds a check to verify that the time in the trace logs
flows as expected (each timestamp must be greater or equal to the
previous one). That should catch a potential misorder of the trace
entries and also such bogus values as above.