base_record/test_overhead.sh: callgraph tests can be optionally disabled
The following options have been added:

PARAM_RECORD_CALLGRAPH_FP      [default is "y"]
PARAM_RECORD_CALLGRAPH_DWARF   [default is unset]

The values "y" and "n" are clear. When PARAM_RECORD_CALLGRAPH_DWARF is
unset, it should be decided whether to run the test according to a
particular environment.

Currently, the decision is based on arch (DWARF callgraph test is run
only on aarch64 and x86_65), but in future, some more intelligent
feature detection would be nice to have.