base_probe/test_exact_counts.sh: we need bigger sample ring buffer
Sometimes, when perf is not able to process collected samples fast
enough, we may see the following error:

Warning:
Processed 60874 events and lost 8 chunks!
Check IO/CPU overload!

This is not a bug, unless it happens even with large ring buffer.
Thus, this commit increases the ring buffer size in the most edge
case.