base_trace/test_basic.sh: handle unpaired syscall '?' timestamps
Adjusts the test to the consequences of the following upstream patch:

commit fd2b2975149f5f7099693027cece81b16842964a
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed Mar 29 16:37:51 2017 -0300

    perf trace: Handle unpaired raw_syscalls:sys_exit event

Before the patch, perf showed bogus systemcall timestamp values for
syscalls that had started before attaching perf-trace.

After the patch, perf shows questionmarks like the following:

         ? (     ?   ): something/732  ... [continued]: poll()) = 0 Timeout

... which makes more sense, since the time is unknown.

This change makes the testsuite to accept it and also to REQUIRE
it in the perf-trace's process attach test.