base_trace/test_basic.sh: do not require timestamp order in systemwide
Note:
In case the perf-trace's behaviour changes, this may be always reverted.

Current behaviour of perf-trace is that when it hits a syscall, it tries
to wait until it finishes, to print all its information at one place,
but if another syscall comes, the original one is printed as "unfinished"
and when it finishes, it is printed again as "continued", but with the
original timestamp of when it was called. Thus, a mis-order in timestamps
can happen, when there are more than one process traced.

So the following is considered OK:

  225.158 (144.498 ms): something/2535 select() ...
  369.656 ( 0.014 ms): something_else/769 inotify_add_watch() = -1 ENOENT
  225.158 (1001.059 ms): something/2535  ... [continued]: select()) = 0 Timeout