base_stat/test_intel_uncore.sh: allow some cbox events to be unsupported
On some machines, such as Intel Sandy Bridge or Ivy Bridge, it happens
that perf offers more cbox events than it actually supports. This is OK,
perf just lists 4 cbox domains and the CPU has only one planted.

With this patch, the test "waives" potential failures of cbox events,
in case there has been already one cbox event passing (cbox_0).

Note:
It is possible that the only working cbox would be cbox_1 and in such
situation the test would fail on not-working cbox_0. If it happens, we
will probably need a more generic fix here.