base_stat/test_advanced_options.sh: adjusted threshold
The idea of the test is that measuring the second half of a command
should give us roughly half event counts. This would be true if the
events were uniformly distributed. The real count is less than half
because the beginning does the biggest part of the work.  The lower
threshold is necessary too - to be sure that the values make sense.
On some aarch64 machine, I found out that the "rest" is pretty much
constant (probably just clean-up), so 0.2 was too high. Sleep isn't
a good load for this experiment...