base_record/test_overhead: fix the load-killing assert
The previous version was too strict and useless. We want the load to
be killed and we assert this to be aware if it's not. However, the
previous implementation was failing a lot because the following logs
don't always get into the log file, but they are printed to console.

[...]
./test_overhead.sh: line 141: 97177 Terminated              $CURRENT_TEST_DIR/examples/load 99999 > /dev/null
[...]

There is no reason why the testsuite should fail due to this.

In case the current implementation does not help with this, it'd
be best to remove the asserts from there completely and kill the
load just "best-effort".