base_mem/test_addresses.sh: new testcase for --phys-data
There are two switches in perf-record: --data and --phys-data which
should make samples contain memory addresses of data touched by the
instruction running during the sample.

The --data option captures the virtual address, the --phys-data opt
leads to its translation to physical address.

Physical address translation works with mem-events. However, it can
lead to 0x0,  which happens either when some bogus address is used,
or when the translation is not finished within an atomic frame. Due
to that we must allow 0x0 there.