base_stat/test_pmu.sh: new test for the "Kernel PMU events"
Some of the "hardware" events provided by CPU hardware PMU counters
have always been available in perf as "hw" or "hwcache" events. The
hw and hwcache event sets were always very limited compared to what
CPUs usually offer.  This is  because the aim was to cover the most
common events only that should be available +- everywhere.

The full eventset offered by the CPU (-and OProfile) was accessible
in perf too, via "rUUEE" or "cpu/event=0xEE,umask=0xUU/", where the
"EE" specified event number and "UU" unit mask.

Nowadays,  there is an approach to have all the available events in
perf too, they're specified in JSON files along with unit masks and
other modifiers in tools/perf/pmu-events tree. During a build, JSON
files are parsed and converted into C files,  so all the data is in
the end compiled into perf binary.

The testcases added in this commit should sanity-test these events.
Since this is very dependent on the target's CPU, thus pretty prone
to fail and also since it may take long to execute such test, there
is a parameter PARAM_STAT_ALL_PMU_EVENTS that disables it.