base_stat/test_pmu.sh: compatibility fix
Some PMU events have '/' in their names, so we have to replace
it by '_' in the log file names.

Also, stderr is redirected to /dev/null, since perf is now giving
advices about proper event usage, such as:

Add unc_p_clockticks event to groups to get metric expression for unc_p_freq_ge_2000mhz_cycles

This is nice for users, but not so easy to just handle by the test
so it would just follow the advices. So for now, let's just ignore
them to keep the screen clean. The exitcode assert should make sure
no error is hidden.