common/_verify_patterns_.sh: added pattern file verificator
In the common/patterns.sh, there have always been examples of what
the regexps should accept. Let's use them to verify the regexps in
the file, so an update of the file can be easily testable.

A negative example type has been added too, to verify that regexps
refuse what should be refused, like the following:

  export RE_PATH_ABSOLUTE="(?:\/$RE_FILE_NAME)+"
  # A full filepath
  # Examples:
  #    /usr/lib64/somelib.so.5.4.0
  #    /lib/modules/4.3.0-rc5/kernel/fs/xfs/xfs.ko
  #    /usr/bin/mv
  #!   some/relative/path
  #!   ./some/relative/path