base_stat/test_pmu.sh: disabling hv_* ppc events and parametrized events
Parametrized events need special handling:

  hv_24x7/CPM_0THRD_NON_IDLE_PCYC,domain=?,core=?/

Something needs to be put instead of the questionmarks, the test needs
to know what. This is solved by separate test test_powerpc_hv_24x7.sh.

Some hv_* events are not parametrized, but they cause troubles too. No
matter that this one does not have any questionmark:

  hv_gpci/system_tlbie_count_and_time_time_spent_issuing_tlbies/

Since in this testcase, there is not even an attempt to detect whether
hv_* stuff should work or not, it's better to skip these.

The "big grep" will probably need to be replaced by a nicer solution
soon.