base_record/test_evlist.sh: reset the max_sample_rate in loop
It resulted necessary to check/reset the max_sample_rate each
time in the loop, since if it's done only once before the loop,
the first loop pass (-F 100) may cause the kernel to lower the
max_sample_rate again, so then the other tests may fail due to
that.

It does not happen often, but I have encountered a machine where
max_sample_rate gets regularly lowered to 600, which causes the
F=1000 test to fail.