base_record/test_intel_pt.sh: New test for Intel Processor Tracing
This test tests Intel Processor Trace functionality. The event intel_pt//
available on newer Intel CPUs (e.g. Skylake) is able to trace all function
calls a program issues during its execution. The example examples/load has
a function_F() that calls function_a() and function_b(). The number of calls
of each function should be equal to number of returns (of course) and equal
for both of te functions. It is also true that it should be all equal to
a sum of an arithmetic row from 1 to N-1 for any N (should be 55 for N=11).
Using this, we can verify that the intel_pt// event actually traces all
the function calls and does not miss any.