base_record/test_intel_pt.sh: bugfix + adapt to new format
1) Fixing a typo in the load execution. The "$NO" variable was empty,
thus, the default 300 was used. That leads to too big files.

2) With current kernel/perf version, the perf-script --itrace output
can contain function offsets too:

So instead of:
7f4687f58cfe strcmp (/usr/lib64/ld-2.17.so) =>  7f4687f58cf0 strcmp (/usr/lib64/ld-2.17.so)

We now have:
7f4687f58cfe strcmp+0x2f (/usr/lib64/ld-2.17.so) =>  7f4687f58cf0 strcmp+0x21 (/usr/lib64/ld-2.17.so)

When the functions are the same, the line means some branching within
the function itself. When the function names differ, the function at
the left calls the function at the right. Thus, the right function's
offset should be +0x0.