base_stat/test_record_report.sh: fixed record/report output inconsistency
While both `perf stat` and `perf stat record` print system/user time info
(like the following), `perf stat report` does not print it. I've tried to
force it to print the stats to,but since they are not stored in perf.data
file, it would print invalid values instead. Thus, to satisfy the diff, I
need to cut it off from the original.