base_record/test_basic.sh: adding a test to cover -k mono crash
The `perf record -k mono -- true` command used to crash when the
output is redirected to stdout. This test reproduces the behavior.

Perf record usually stores the data output to perf.data file. It
also can be redirected to stdout by '-o -' or, surprisingly, by
the simple fact that perf's stdout stream is piped!!!

So, `perf record something | cat` causes perf record to put the
binary data to stdout!

However, when adding '-k mono', it used to crash, which is fixed
by the following patch:

commit c952b35f4b15dd1b83e952718dec3307256383ef
Author: Song Liu <songliubraving@fb.com>
Date:   Wed Jun 19 18:04:53 2019 -0700

    perf header: Assign proper ff->ph in perf_event__synthesize_features()

This testcase is here to cover this scenario.