base_record/test_intr_regs.sh: fix failure on some KVM guests
On some KVM guests this testcase used to fail. The reason is pretty
simple: If the guest does not support HW events (the default event
does not work), perf record fallbacks to "cpu-clock" as a back-up
default event. This is fine and it allows various perf-record tests
pass even on environments without PMU.

However, the previous load ($CMD_SIMPLE which equals to 'true') does
not provide enough samples for cpu-clock (while providing enough
samples for cycles), it used to get 0-1 samples, so sometimes, there
were 0 samples, so the test failed.

This patch changes the load to something a bit heavier, so it should
not be on the edge anymore.