base_stat/test_metrics.sh: New test
Adding a testcase for testing perf-stat with "metrics". Metrics are
sets of multiple events put together to mean something meaningful.
For example a metric called "cpi" counts automatically instructions
and cycles and calculates cycles-per-instruction ratio. This is in
perf to help users so they could gain more from the large sets of
PMU events offered by modern CPUs, which no one except the designers
of the CPU really knows what they are for.

This test is disabled by default to not screw up CI, since it still
has quite lot of fails.