common: implementing a new depth-level runmode mechanism
With this, testsuite could be run in one of three modes: BASIC, STANDARD
and EXPERIMENTAL. The modes are defined in common/settings.sh. The default
mode is STANDARD, which is intended to be used for normal testing.

BASIC mode is intended to be used in CI where you don't want to see fails
caused by unstable tests. Thus, tests that are known trouble-makers can
be marked at the begining by "consider_skipping $RUMNODE_STANDARD" and they
will be skipped if the suite runs in RUNMODE_BASIC (which is lower than
RUNMODE_STANDARD).

EXPERIMENTAL mode is intended to be used for experimental purposes. This
will allow us to add almost-broken unstable testcases for experimenting
and not breaking CI or standard testing.

Tests marked as "consider_skipping $RUNMODE_EXPERIMENTAL" will be skipped
in BASIC or STANDARD runmode and run only if the testsuite itself runs in
the EXPERIMENTAL mode.

If a testcase is not marker, it is the same as being marked as BASIC.

Parametrization scripts were divided too, since the runmodes affect the
default parameter settings too. All parametrization* config scripts are
designed to set something "if not set before", so they can be combined
together. Current parameter handling is that the highest priority has
what the user exports, then the parametrization.<number>.sh file that
defines the not-yet-set parameters according to the runmode and then
there is the default parametrization.sh script that has the lowest prio
in the whole stack and defines the parameters that haven't been set yet,
either by a user or by the runmode-dependent parametrization script.