base_probe/test_mixed_tp.sh: fixed testcase fails
1) The "examples/advanced" load had to be replaced by "examples/load",
since the original one produced too little samples, so sometimes it had
zero cpu-clock samples, so the check of "mixed" perf-script output failed.
With the "examples/load" (which causes more CPU load), the number of the
cpu-clock samples should be sufficient.

2) Do not require cpu ("[001]") column in perf-script. Since samples for
'cpu-clock' event do not have CPU attribute set, it might be missing even
for the probe, since the events are grouped together.