base_stat/test_endless_wait.sh: disable in BASIC mode
... and keep in STANDARD and EXPERIMENTAL modes.

This test may fail sometimes due to some lag in process starting or its
termination. The test checks whether the 'time' output contains 2 secs
of real time (which is default for CMD_LONGER_SLEEP) and does not contain
4 secs (which is default for CMD_DOUBLE_LONGER_SLEEP). Sometimes, it
takes 3.001 secs which does not match the "^real\s+0m2" regexp. A bit
slower, but still OK. The bug is reproduced if perf hangs and is killed
after the DOUBLE_LONGER_SLEEP interval.

Increasing CMD_LONGER_SLEEP and CMD_DOUBLE_LONGER_SLEEP would make the
test a bit more stable, however it would make the testsuite slower, since
it is used at many places. Also note that it wouldn't eliminate the above
mentioned problem, since it might always take one sec more than we have
in CMD_LONGER_SLEEP, so the fault is in the primitive regexp matching
used for time checks. For proper solving, we would have to implement some
smarter time checker with thresholds. I don't want to do that, since it
is never 100% reliable anyway.

The best way seems to be to disable the test in CI-like environments.