base_probe/test_listing.sh: Fixed line-listing verification
Before, we relied on the fact that function entry/exit points are
represented as '{' and '}' in perf, so it looked like this:

  $ perf probe -L vfs_read
  <vfs_read@/usr/src/debug/kernel-3.10.0-1062.1.2.el7/linux-3.10.0-1062.1.2.el7.x86_64/fs/read_write.c:0>
        0  ssize_t vfs_read(struct file *file, char __user *buf, size_t count, loff_t *pos)
-->     1  {
                  ssize_t ret;

        4         if (!(file->f_mode & FMODE_READ))
  [...]

(above, the line with '{' can be probed, so it has a number)

Now, the lines with only '{' can't be probed (and it probably does
not make sense, since it is not a valid command), so they don't have
numbers assigned in the perf listing:

  $ perf probe -L vfs_read
  <vfs_read@/usr/src/debug/kernel-4.18.0-167.el8/linux-4.18.0-167.el8.x86_64/fs/read_write.c:0>
        0  ssize_t vfs_read(struct file *file, char __user *buf, size_t count, loff_t *pos)
-->        {
        2         ssize_t ret;
  [...]

Removing the regular expressions that want it then.

In order to not lame the checks that much, adding some new regular
expressions. Let's try to rely on having if-conditions checking the
very common 'ret' variable...